Change to type float

+++ src/Activations/activations.cpp
@@ -1,6 +1,5 @@
-#include <Eigen/Dense>
 #include <cmath>
 
-double sigmoid(const double& input) {
+float sigmoid(const float& input) {
     return 1.0 / (1.0 + exp(-input));
 };
